requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.5
flatbuffers==25.9.23
h11==0.16.0
idna==3.11
jh2==5.0.10
niquests==3.18.1
numpy==2.4.3
pandas==3.0.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
qh3==1.6.0
requests==2.32.5
six==1.17.0
urllib3==2.6.3
urllib3-future==2.17.901
wassima==2.0.5
